Update to support BSD man(1) as well as GNU man(1)
BSD man(1) doesn't support the same `--sections` format an GNU does.
GNU man(1) does support the BSD version of the argument, though,
so it is compatible to use `-S 1:8` with both.

This switches to the BSD version, `-S`, and colon-separated section
numbers.  Tested on macOS and GNU/Linux, where it works correctly.